Changed the data type of a b c and did some code review/reading

+++ src/validate/validate_line.h
@@ -4,4 +4,4 @@ int is_empty(char * line);
 
 int is_number(char * result);
 
-int check_validation(char * line, int n, double * a, double *b, double *c);
+int check_validation(char * line, int n, float * a, float * b, float * c);
